admin roles update

+++ resources/views/livewire/admin/department/department.blade.php
@@ -28,7 +28,9 @@
             <div class="tab-pane fade @if($active == 'it') show active @endif " id="IT-management-tab">
                 <div class="container-fluid">
                     <div class="table-responsive">
-                        <button class="btn btn-success float-right mt-2 mb-2" wire:click="add_q_and_a('IT')">Add IT Q&A</button>
+                        @if( $access_role['C'] == 1)
+                            <button class="btn btn-success float-right mt-2 mb-2" wire:click="add_q_and_a('IT')">Add IT Q&A</button>
+                        @endif
                         <table id="example1" class="table table-hover table-bordered" style="width:100%">
                             <caption>These data can be modified and structured before calling as a function in the model</caption>
                             <thead class="thead-dark">
@@ -83,8 +85,12 @@
                                         @endforeach
                                         </td>
                                         <td class="text-center">
-                                            <button class="btn btn-primary" wire:click="edit_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-edit'></i></button>
+                                            @if( $access_role['U'] == 1)
+                                                <button class="btn btn-primary" wire:click="edit_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-edit'></i></button>
+                                            @endif
+                                            @if( $access_role['D'] == 1)
                                             <button class="btn btn-danger" wire:click="delete_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-trash'></i></button>
+                                            @endif
                                         </td>
                                     </tr>
                                 @empty
@@ -101,7 +107,9 @@
             <div class="tab-pane fade @if($active == 'cs') show active @endif " id="CS-management-tab">
                 <div class="container-fluid">
                     <div class="table-responsive">
-                        <button class="btn btn-success float-right mt-2 mb-2" wire:click="add_q_and_a('CS')">Add CS Q&A</button>
+                        @if( $access_role['C'] == 1)
+                            <button class="btn btn-success float-right mt-2 mb-2" wire:click="add_q_and_a('CS')">Add CS Q&A</button>
+                        @endif
                         <table id="example2" class="table table-hover table-bordered" style="width:100%">
                             <caption>These data can be modified and structured before calling as a function in the model</caption>
                             <thead class="thead-dark">
@@ -157,8 +165,12 @@
                                         @endforeach
                                         </td>
                                         <td class="text-center">
-                                            <button class="btn btn-primary" wire:click="edit_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-edit'></i></button>
-                                            <button class="btn btn-danger" wire:click="delete_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-trash'></i></button>
+                                            @if( $access_role['U'] == 1)
+                                                <button class="btn btn-primary" wire:click="edit_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-edit'></i></button>
+                                            @endif
+                                            @if( $access_role['D'] == 1)
+                                                <button class="btn btn-danger" wire:click="delete_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-trash'></i></button>
+                                            @endif 
                                         </td>
                                     </tr>
                                 @empty
@@ -173,7 +185,9 @@
             <div class="tab-pane fade @if($active == 'act') show active @endif " id="ACT-management-tab">
                 <div class="container-fluid">
                     <div class="table-responsive">
-                        <button class="btn btn-success float-right mt-2 mb-2" wire:click="add_q_and_a('ACT')">Add ACT Q&A</button>
+                        @if( $access_role['C'] == 1)
+                            <button class="btn btn-success float-right mt-2 mb-2" wire:click="add_q_and_a('ACT')">Add ACT Q&A</button>
+                        @endif
                         <table id="example2" class="table table-hover table-bordered" style="width:100%">
                             <caption>These data can be modified and structured before calling as a function in the model</caption>
                             <thead class="thead-dark">
@@ -229,8 +243,12 @@
                                         @endforeach
                                         </td>
                                         <td class="text-center">
-                                            <button class="btn btn-primary" wire:click="edit_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-edit'></i></button>
-                                            <button class="btn btn-danger" wire:click="delete_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-trash'></i></button>
+                                            @if( $access_role['U'] == 1)
+                                                <button class="btn btn-primary" wire:click="edit_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-edit'></i></button>
+                                            @endif
+                                            @if( $access_role['D'] == 1)
+                                                <button class="btn btn-danger" wire:click="delete_q_and_a({{$value['q_and_a_id']}})"><i class='bx bxs-trash'></i></button>
+                                            @endif
                                         </td>
                                     </tr>
                                 @empty
